Folder structure modified
