+++ http_test.py
@@ -12,6 +12,8 @@ response = requests.put(BASE + "graph", {
     "goal": 4,
     "edges": "0-1,0-2,0-3,3-4"
 })
+
+print(response.json())
 '''
 
 '''
@@ -27,7 +29,30 @@ print(response.json())
 
 '''
 Para conjuntos disjuntos
+
+response = requests.put(BASE + "sets", {
+    "final_node": 9,
+    "edges": "0-1,0-2,1-2,4-5,4-6,5-6,7-8,7-9,8-9"
+})
+
+print(response.json())
 '''
+response = requests.put(BASE + "graph", {
+    "start": 0,
+    "goal": 4,
+    "edges": "0-1,0-2,0-3,3-4"
+})
+
+print(response.json())
+
+
+response = requests.put(BASE + "priority_queue", {
+    "names": "Melissa,Juan,Ernesto,Andres,Javier,Vanessa,Mariana,Pedro,Luz",
+    "popularity": "90,87,10,24,37,65,70,14,93",
+    "times_spoken": "15,30,5,1,55,19,25,173,47"
+})
+print(response.json())
+
 response = requests.put(BASE + "sets", {
     "final_node": 9,
     "edges": "0-1,0-2,1-2,4-5,4-6,5-6,7-8,7-9,8-9"
